requirements.txt
certifi==2020.4.5.1
click==7.1.2
cycler==0.10.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
Werkzeug==1.0.1
